fix buuuuuuuuuuuugggggggggggggg`zzzzzzzzzz

+++ vm/op_live.c
@@ -12,7 +12,7 @@ void live(void)
 	g_bogies->last_breath = g_arena->round;
 	g_bogies->live_op++;
 	if (g_flags->v == 1 || g_flags->v == 30)
-		ft_printf("P %d | live %d\n", g_bogies->num, arg);
+		ft_printf("P %4d | live %d\n", g_bogies->num, arg);
 	if (arg >= g_arena->champ_num * -1  && arg <= -1)
 		g_arena->last_stand = arg;
 	g_arena->mortal_flip++;
@@ -21,14 +21,14 @@ void live(void)
 
 void zjump(void)
 {
-	int				arg;
+	short	arg;
 
 	if (g_bogies->its_a_highnoon != g_arena->round)
 		return;
 	arg = get_tdir_small_size(g_bogies->index + 1);
 	arg %= IDX_MOD;
 	if (g_flags->v == 1 || g_flags->v == 30)
-		ft_printf("P %d | zjmp %d %s\n", g_bogies->num, arg, g_bogies->carry ? "OK" : "FAILED");
+		ft_printf("P %4d | zjmp %d %s\n", g_bogies->num, arg, g_bogies->carry ? "OK" : "FAILED");
 	if (g_bogies->carry)
 		move_caret(arg - 1);
 	else
